requirements.txt
anyio==3.3.4
asgiref==3.4.1
attrs==19.3.0
Automat==0.8.0
blinker==1.4
bs4==0.0.1
cached-property==1.5.2
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.0.7
Click==7.0
colorama==0.4.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
distro==1.4.0
dnspython==2.1.0
email-validator==1.1.3
entrypoints==0.3
fastapi==0.70.0
graphql-core==3.1.6
greenlet==1.1.2
h11==0.12.0
httplib2==0.14.0
httptools==0.2.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==3.0.2
jose==1.0.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.27
netifaces==0.10.4
oauthlib==3.1.0
orjson==3.6.4
packaging==21.2
passlib==1.7.4
pipreqs==0.4.11
pluggy==1.0.0
protobuf==3.19.1
psycopg2-binary==2.9.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pydantic==1.8.2
Pygments==2.10.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyrsistent==0.15.5
pyserial==3.4
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.1
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.3.1
requests==2.26.0
requests-unixsocket==0.2.0
rsa==4.7.2
SecretStorage==2.3.1
sentinel==0.3.0
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sniffio==1.2.0
soupsieve==2.2.1
SQLAlchemy==1.4.26
ssh-import-id==5.10
starlette==0.16.0
strawberry-graphql==0.85.1
toml==0.10.2
typing-extensions==3.10.0.2
ujson==4.2.0
urllib3==1.25.8
uvicorn==0.15.0
uvloop==0.16.0
wadllib==1.3.3
watchgod==0.7
websockets==10.0
Werkzeug==1.0.1
yarg==0.1.9
zipp==1.0.0
